requirements.txt
django
djangorestframework
psycopg2-binary
pyyaml
ujson
requests
djoser
django_filter